Re-initialize repo.
I had added a bunch of PDFs and PNGs to the repo without .gitignoring
them, and at 28MB, the objects were just a little ridiculous. Rather
than trying to go through `git filter-branch for all of the files, and
because the repo is so young, I decided to re-start the history.